Move types definied in the settlement pallet to primitives; Move Event and WightInfo to common; Move schedule_name_prefix constants to primitives
