+++ pallets/runtime/tests/src/sto_test.rs
@@ -1,32 +1,29 @@
-use super::{
-    asset_test::{allow_all_transfers, max_len_bytes, set_timestamp},
-    exec_noop, exec_ok,
-    storage::{make_account_with_portfolio, TestStorage, User},
-    ExtBuilder,
-};
-use pallet_asset as asset;
-use pallet_compliance_manager as compliance_manager;
-use pallet_settlement::{InstructionStatus, VenueDetails, VenueId, VenueType};
+use frame_support::{assert_noop, assert_ok};
+use sp_runtime::DispatchError;
+
 use pallet_sto::{
     Fundraiser, FundraiserId, FundraiserName, FundraiserStatus, FundraiserTier, PriceTier,
     MAX_TIERS,
 };
+use polymesh_primitives::settlement::{InstructionStatus, VenueDetails, VenueId, VenueType};
 use polymesh_primitives::{
     asset::AssetType, checked_inc::CheckedInc, PortfolioId, Ticker, WeightMeter,
 };
-
-use crate::storage::provide_scope_claim_to_multiple_parties;
-use frame_support::{assert_noop, assert_ok};
-use sp_runtime::DispatchError;
 use test_client::AccountKeyring;
 
+use super::asset_test::{allow_all_transfers, max_len_bytes, set_timestamp};
+use super::storage::{
+    make_account_with_portfolio, provide_scope_claim_to_multiple_parties, TestStorage, User,
+};
+use super::{exec_noop, exec_ok, ExtBuilder};
+
 type Origin = <TestStorage as frame_system::Config>::RuntimeOrigin;
-type Asset = asset::Module<TestStorage>;
+type Asset = pallet_asset::Module<TestStorage>;
 type Sto = pallet_sto::Module<TestStorage>;
 type Error = pallet_sto::Error<TestStorage>;
 type EAError = pallet_external_agents::Error<TestStorage>;
 type PortfolioError = pallet_portfolio::Error<TestStorage>;
-type ComplianceManager = compliance_manager::Module<TestStorage>;
+type ComplianceManager = pallet_compliance_manager::Module<TestStorage>;
 type Settlement = pallet_settlement::Module<TestStorage>;
 type Timestamp = pallet_timestamp::Pallet<TestStorage>;
 type System = frame_system::Pallet<TestStorage>;
