+++ pallets/settlement/Cargo.toml
@@ -15,7 +15,6 @@ pallet-external-agents = { path = "../external-agents", default-features = false
 polymesh-primitives = { path = "../../primitives", default-features = false }
 polymesh-primitives-derive = { path = "../../primitives_derive", default-features = false }
 pallet-compliance-manager = { path = "../compliance-manager", default-features = false }
-#polymesh-contracts = { path = "../contracts", default-features = false }
 pallet-statistics = { path = "../statistics", default-features = false }
 pallet-nft = { path = "../nft", default-features = false }
 
@@ -36,7 +35,6 @@ sp-api = { version = "4.0.0-dev", default-features = false }
 frame-system = { version = "4.0.0-dev", default-features = false }
 frame-support = { version = "4.0.0-dev", default-features = false }
 pallet-timestamp = { version = "4.0.0-dev", default-features = false }
-#pallet-contracts = { version = "4.0.0-dev", default-features = false }
 pallet-scheduler = { version = "4.0.0-dev", default-features = false }
 
 # Only in STD
